requirements.txt
pandas
bs4
openpyxl
requests
python-dotenv
mysql-connector-python
sqlalchemy
matplotlib
